Term Paper
A comprehensive solution, based on the Yii2 basic
and advanced frameworks, shall be presented in this report to
manage tour packages, purchase tour packages or make
payment of tour packages. In addition, it is discussed how to
integrate ownCloud into the efficient sharing of files. Increasing
demand for complete tour package management, and the need
to process payments safely and efficiently, is at the heart of this
work. The lack of a common system combining package
management, reservation, payment processing and file sharing
functionalities is addressed. In order to address this problem, we
have proposed a methodology that uses the Yii2 frameworks'
ability to create high performance and scalable applications. In
order to achieve effective file sharing and collaboration, the
system provides a number of features including package
creation, booking management, secure online payment channels
as well as seamless integration with ownCloud. The results show
improvements in efficiency, enhanced safety and simplification
of procedures related to package management, booking,
payment and file sharing. The overall aim of this solution is to
improve the operation of tour agencies and clients in general by
providing a comprehensive and coherent platform.